refactor: move the package
- model, view, controller
