+++ src/main/kotlin/lotto/view/LottoTicketDto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.view
 
 data class LottoTicketDto(val numbers: List<Int>) {
 }
