+++ src/test/kotlin/lotto/GameTest.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.model.Game
+import lotto.model.WinningLotto
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
