+++ src/test/kotlin/lotto/WinningLottoTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.model.WinningLotto
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
